Fixing the way temporary interfaces are set. This broke copying of blobs
Originally the code removed all directly provided interfaces of an object
when the CopiedEvent was thrown. It also copied the original object from
the event, and added it to the object.
It later tried to restore the interfaces by adding the interfaces from
the original object.
When copying folders, plone.app.container triggers all the events for its
children too. It reuses the event object.
As plone.app.relations copies the object from the event, things get interesting
when the par event handler handles the event for any of its subobjects.
It removes all directly provided interfaces from a blob object, and
in a later step it adds all directly provided interfaces from the parent
folder. Also, if somebody registers an event handler inbetween that tries
to get the primary field from a blob, it won't get the primary field,
because plone.app.blob relies on the temporary removed interfaces to
get the right primary field.

This fixes #10555

svn path=/plone.app.relations/trunk/; revision=36645